requirements.txt
# Flask and Related Extensions
Flask==2.3.3               # Core framework
Flask-Cors==3.0.10         # Cross-Origin Resource Sharing
Flask-Restx==1.1.0         # API Documentation and Routing
Flask-SQLAlchemy==3.0.5    # SQLAlchemy ORM for Flask
Flask-Migrate==4.0.4       # Migrations with Alembic
Flask-JWT-Extended==4.4.4  # JWT authentication

# Database Drivers
psycopg2-binary==2.9.7     # PostgreSQL driver
pyodbc==4.0.39             # SQL Server driver

# Environment Management
python-dotenv==1.0.0       # Environment variable management

# Security
cryptography==41.0.1       # Encryption for token storage

# Task Queue
celery[redis]==5.3.1       # Task queue with Redis
celery[beat]==5.3.1        # Periodic task scheduling

# Testing
pytest==7.4.2              # Testing framework
pytest-mock==3.11.1        # Mocking in tests
